* add some example JDL and its generated JPA counter part
* fleshing out the tests

* improved docs and readme
