+++ build.gradle
@@ -23,10 +23,10 @@ dependencies {
     compile group: 'org.hibernate', name: 'hibernate-validator', version:'5.2.4.Final'
     runtime group: 'ch.qos.logback', name: 'logback-classic', version:'1.1.5'
     testCompile group: 'junit', name: 'junit', version:'4.12'
-    testCompile "org.springframework.boot:spring-boot-starter-data-jpa:1.5.4.RELEASE"
-    testCompile "org.hibernate:hibernate-jcache:5.2.8.Final"
-    testCompile "org.hibernate:hibernate-envers:5.2.8.Final"
-    testCompile "com.fasterxml.jackson.core:jackson-annotations:2.8.9"
+    compile "org.springframework.boot:spring-boot-starter-data-jpa:1.5.4.RELEASE"
+    compile "org.hibernate:hibernate-jcache:5.2.8.Final"
+    compile "org.hibernate:hibernate-envers:5.2.8.Final"
+    compile "com.fasterxml.jackson.core:jackson-annotations:2.8.9"
 }
 
 task execute(type: JavaExec, dependsOn: 'prep') {
