+++ src/main/java/io/github/jhipster/jpa2jdl/example/entities/AbstractAuditingEntity.java
@@ -1,4 +1,4 @@
-package io.github.jhipster.jpa2jdl.test.unit.entities;
+package io.github.jhipster.jpa2jdl.example.entities;
 
 import java.io.Serializable;
