+++ src/main/java/io/github/jhipster/jpa2jdl/example/entities/Authority.java
@@ -1,4 +1,4 @@
-package io.github.jhipster.jpa2jdl.test.unit.entities;
+package io.github.jhipster.jpa2jdl.example.entities;
 
 import org.hibernate.annotations.Cache;
 import org.hibernate.annotations.CacheConcurrencyStrategy;
