+++ src/main/java/io/github/jhipster/jpa2jdl/example/entities/Constants.java
@@ -1,4 +1,4 @@
-package io.github.jhipster.jpa2jdl.test.unit.entities;
+package io.github.jhipster.jpa2jdl.example.entities;
 
 /**
  * Application constants.
