--- src/test/java/io/github/jhipster/jpa2jdl/test/unit/entities/bugtracker/Label.java
@@ -1,112 +1,0 @@
-package io.github.jhipster.jpa2jdl.test.unit.entities.bugtracker;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import org.hibernate.annotations.Cache;
-import org.hibernate.annotations.CacheConcurrencyStrategy;
-
-import javax.persistence.*;
-import javax.validation.constraints.*;
-import java.io.Serializable;
-import java.util.HashSet;
-import java.util.Set;
-import java.util.Objects;
-
-/**
- * A Label.
- */
-@Entity
-@Table(name = "label")
-@Cache(usage = CacheConcurrencyStrategy.NONSTRICT_READ_WRITE)
-public class Label implements Serializable {
-
-    private static final long serialVersionUID = 1L;
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "sequenceGenerator")
-    @SequenceGenerator(name = "sequenceGenerator")
-    private Long id;
-
-    @NotNull
-    @Size(min = 3)
-    @Column(name = "jhi_label", nullable = false)
-    private String label;
-
-    @ManyToMany(mappedBy = "labels")
-    @JsonIgnore
-    @Cache(usage = CacheConcurrencyStrategy.NONSTRICT_READ_WRITE)
-    private Set<Ticket> tickets = new HashSet<>();
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getLabel() {
-        return label;
-    }
-
-    public Label label(String label) {
-        this.label = label;
-        return this;
-    }
-
-    public void setLabel(String label) {
-        this.label = label;
-    }
-
-    public Set<Ticket> getTickets() {
-        return tickets;
-    }
-
-    public Label tickets(Set<Ticket> tickets) {
-        this.tickets = tickets;
-        return this;
-    }
-
-    public Label addTicket(Ticket ticket) {
-        this.tickets.add(ticket);
-        ticket.getLabels().add(this);
-        return this;
-    }
-
-    public Label removeTicket(Ticket ticket) {
-        this.tickets.remove(ticket);
-        ticket.getLabels().remove(this);
-        return this;
-    }
-
-    public void setTickets(Set<Ticket> tickets) {
-        this.tickets = tickets;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Label label = (Label) o;
-        if (label.getId() == null || getId() == null) {
-            return false;
-        }
-        return Objects.equals(getId(), label.getId());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(getId());
-    }
-
-    @Override
-    public String toString() {
-        return "Label{" +
-            "id=" + getId() +
-            ", label='" + getLabel() + "'" +
-            "}";
-    }
-}
