+++ src/test/resources/README.md
@@ -1,0 +1,3 @@
+23.06.2017
+
+git cloned from https://github.com/jhipster/jdl-samples
